+++ DemoApp/Sources/Components/Debug/Items/FeatureFlags/DebugMenu+FeatureFlagsView.swift
@@ -8,17 +8,18 @@ import SwiftUI
 extension AppEnvironment {
 
     /// Wraps a debug menu view in an identifiable container.
-    struct FeatureFlag: Identifiable {
+    struct FeatureFlag: Identifiable, @unchecked Sendable {
         var id: UUID = .init()
         var viewProvider: () -> AnyView
     }
 
-    nonisolated(unsafe) static var featureFlags: [FeatureFlag] = [
+    @MainActor static var featureFlags: [FeatureFlag] = [
         .init { .init(DebugMenu.VideoProcessingPipelineToggleView()) },
         .init { .init(DebugMenu.CapturingPipelineToggleView()) },
         .init { .init(DebugMenu.VideoRenderingMenuView()) },
         .init { .init(DebugMenu.CallJoinInterceptorSelector()) },
-        .init { .init(DebugMenu.HighScaleLivestreamPublisherHintToggleView()) }
+        .init { .init(DebugMenu.HighScaleLivestreamPublisherHintToggleView()) },
+        .init { .init(DebugMenu.EncryptionKeyMenuView()) }
     ]
 }