+++ DemoApp/Sources/Extensions/URL+Convenience.swift
@@ -12,8 +12,9 @@ extension URL {
 
     var queryParameters: [String: String] {
         guard
-            let components = URLComponents(url: self, resolvingAgainstBaseURL: true),
-            let queryItems = components.queryItems else { return [:] }
+            let components = URLComponents(url: self, resolvingAgainstBaseURL: false),
+            let queryItems = components.queryItems
+        else { return [:] }
         return queryItems.reduce(into: [String: String]()) { (result, item) in
             result[item.name] = item.value
         }